moving routes to main; moving storage; removing config.py etc.
