+++ backend/ops/transcription_job/constants.py
@@ -1,2 +1,21 @@
+import dataclasses
 
-MAX_JOB_AGE_SECS = 10 * 60
+MAX_JOB_AGE_SECS = 10 * 60 # How long a job can take
+
+@dataclasses.dataclass
+class ModelSpec:
+    name: str
+    params: str
+    relative_speed: int  # Higher is faster
+
+supported_whisper_models = {
+    "tiny.en": ModelSpec(name="tiny.en", params="39M", relative_speed=32),
+    # Takes around 3-10 minutes to transcribe a podcast, depending on length.
+    "base.en": ModelSpec(name="base.en", params="74M", relative_speed=16),
+    "small.en": ModelSpec(name="small.en", params="244M", relative_speed=6),
+    "medium.en": ModelSpec(name="medium.en", params="769M", relative_speed=2),
+    # Very slow. Will take around 45 mins to 1.5 hours to transcribe.
+    "large": ModelSpec(name="large", params="1550M", relative_speed=1),
+}
+
+DEFAULT_MODEL = supported_whisper_models["base.en"]
