+++ backend/ops/transcription_job/operators.py
@@ -8,16 +8,16 @@ from backend.ops.storage import APP_VOLUME, get_episode_metadata_path, get_podca
 from backend.src.podcast.functions.download import store_original_audio
 from backend.src.transcription_job.functions.silences import split_silences
 
-from backend.ops.transcription_job.constants import MAX_JOB_AGE_SECS
+from backend.ops.transcription_job.constants import ModelSpec
 from backend.src.podcast.types import EpisodeMetadata
 
-from backend import config
+from backend.ops import storage
 
-logger = config.get_logger(__name__)
+from backend import _utils; logger = _utils.get_logger(__name__)
 
 @stub.function(
     image=APP_IMAGE,
-    network_file_systems={config.CACHE_DIR: APP_VOLUME},
+    network_file_systems={storage.CACHE_DIR: APP_VOLUME},
     cpu=4,
     # gpu="any",
     timeout=600,
@@ -26,7 +26,7 @@ def transcribe_segment(
     start: float,
     end: float,
     audio_filepath: pathlib.Path,
-    model: config.ModelSpec,
+    model: ModelSpec,
 ):
     import tempfile
     import time
@@ -48,7 +48,7 @@ def transcribe_segment(
         use_gpu = torch.cuda.is_available()
         device = "cuda" if use_gpu else "cpu"
         model = whisper.load_model(
-            model.name, device=device, download_root=config.MODEL_DIR
+            model.name, device=device, download_root=storage.MODEL_DIR
         )
         result = model.transcribe(f.name, language="en", fp16=use_gpu)  # type: ignore
 
@@ -66,13 +66,13 @@ def transcribe_segment(
 
 @stub.function(
     image=APP_IMAGE,
-    network_file_systems={config.CACHE_DIR: APP_VOLUME},
+    network_file_systems={storage.CACHE_DIR: APP_VOLUME},
     timeout=900,
 )
 def transcribe_episode(
     audio_filepath: pathlib.Path,
     result_path: pathlib.Path,
-    model: config.ModelSpec,
+    model: ModelSpec,
 ):
     segment_gen = split_silences(str(audio_filepath))
 
@@ -97,7 +97,7 @@ def transcribe_episode(
 
 @stub.function(
     image=APP_IMAGE,
-    network_file_systems={config.CACHE_DIR: APP_VOLUME},
+    network_file_systems={storage.CACHE_DIR: APP_VOLUME},
     timeout=900,
 )
 def process_episode(podcast_id: str, episode_number: str):
@@ -107,11 +107,11 @@ def process_episode(podcast_id: str, episode_number: str):
     try:
         # pre-download the model to the cache path, because the _download fn is not
         # thread-safe.
-        model = config.DEFAULT_MODEL
-        whisper._download(whisper._MODELS[model.name], config.MODEL_DIR, False)
+        model = storage.DEFAULT_MODEL
+        whisper._download(whisper._MODELS[model.name], storage.MODEL_DIR, False)
 
-        config.RAW_AUDIO_DIR.mkdir(parents=True, exist_ok=True)
-        config.TRANSCRIPTIONS_DIR.mkdir(parents=True, exist_ok=True)
+        storage.RAW_AUDIO_DIR.mkdir(parents=True, exist_ok=True)
+        storage.TRANSCRIPTIONS_DIR.mkdir(parents=True, exist_ok=True)
 
         metadata_path = get_episode_metadata_path(podcast_id)
         with open(metadata_path, "r") as f:
@@ -122,7 +122,7 @@ def process_episode(podcast_id: str, episode_number: str):
                 data_class=EpisodeMetadata, data=data
             )
 
-        destination_path = config.RAW_AUDIO_DIR / episode.guid_hash
+        destination_path = storage.RAW_AUDIO_DIR / episode.guid_hash
         store_original_audio(
             url=episode.audio_download_url,
             destination=destination_path,
@@ -181,8 +181,8 @@ def download_step(fixed_podcast_url): # 1st step
         content_type=response.headers["content-type"],
     )
     # storing the audio intermediately
-    config.RAW_AUDIO_DIR.mkdir(parents=True, exist_ok=True) # must create directory on remote
-    audio_downloaded_path = config.RAW_AUDIO_DIR / "fixed_podcast_audio"
+    storage.RAW_AUDIO_DIR.mkdir(parents=True, exist_ok=True) # must create directory on remote
+    audio_downloaded_path = storage.RAW_AUDIO_DIR / "fixed_podcast_audio"
     humanized_bytes_str = sizeof_fmt(num=len(podcast_download_result.data))
     size_of_download_msg = f"Downloaded {humanized_bytes_str} episode from URL."
     logger.info(size_of_download_msg)
@@ -198,7 +198,7 @@ def download_step(fixed_podcast_url): # 1st step
 
 @stub.function(
     image=APP_IMAGE,
-    network_file_systems={config.CACHE_DIR: APP_VOLUME},
+    network_file_systems={storage.CACHE_DIR: APP_VOLUME},
     timeout=6000,
     retries=10 # 100mins
 )
@@ -245,8 +245,8 @@ def download_transcribe_annotate_fixed_podcast_in_single_job_process(fixed_podca
     logger.info(transcribe_time_taken_msg)
 
     # storing the transcription
-    config.TRANSCRIPTIONS_DIR.mkdir(parents=True, exist_ok=True)
-    transcription_storage_path = config.TRANSCRIPTIONS_DIR / "fixed_podcast_transcript"
+    storage.TRANSCRIPTIONS_DIR.mkdir(parents=True, exist_ok=True)
+    transcription_storage_path = storage.TRANSCRIPTIONS_DIR / "fixed_podcast_transcript"
     with open(transcription_storage_path, "w") as f:
         json.dump(transcription, f, indent=4)
     transcript_json_storage_location_msg = f"Stored transcript at {transcription_storage_path}."
